+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Remove keyboard_support setting (#386)
 * Correctly handle opening of /dev/input/event|jsdev with writing flag
 * Fix pitch for OpenAL
+* Fix some labels in the input editor being truncated
 
 ## [1.4.1] - 2021-01-02
 ### Added